Tareas Cortas/TareaCorta1/flask/Images/flaskMariadbGalera/app/requirement.txt
Flask
mysql-connector-python==8.0.26
